Revert "chore(backend): Moved shared module to util/shared and drop unused"
This reverts commit 2f6a6ac9
