+++ packages/backend/src/shared/index.ts
@@ -1,10 +1,13 @@
 export {
   addClerkPrefix,
+  buildPublishableKey,
   callWithRetry,
   getClerkJsMajorVersionOrTag,
   getScriptUrl,
   isDevelopmentFromApiKey,
   isProductionFromApiKey,
+  isPublishableKey,
+  isStaging,
   parsePublishableKey,
   deprecated,
   deprecatedProperty,
