+++ packages/backend/src/tokens/interstitial.ts
@@ -4,7 +4,6 @@ import { API_VERSION, USER_AGENT } from '../constants';
 // DO NOT CHANGE: Runtime needs to be imported as a default export so that we can stub its dependencies with Sinon.js
 // For more information refer to https://sinonjs.org/how-to/stub-dependency/
 import runtime from '../runtime';
-import { joinPaths } from '../util/path';
 import {
   addClerkPrefix,
   callWithRetry,
@@ -13,7 +12,8 @@ import {
   getScriptUrl,
   isDevOrStagingUrl,
   parsePublishableKey,
-} from '../util/shared';
+} from '../shared';
+import { joinPaths } from '../util/path';
 import { TokenVerificationError, TokenVerificationErrorAction, TokenVerificationErrorReason } from './errors';
 import type { DebugRequestSate } from './request';
