+++ playground/app-router/package.json
@@ -20,7 +20,7 @@
     "@types/node": "18.16.0",
     "@types/react": "18.0.38",
     "@types/react-dom": "18.0.11",
-    "next": "^13.4.4",
+    "next": "^13.5.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "5.0.4"
